Refactor of UI complete. Down to 30 lines
